Updating logic to fix Order Page Details view CH-40999
