+++ Block/Adminhtml/Container.php
@@ -42,6 +42,11 @@ class Container extends Template
     public $config;
 
     /**
+     * @var Url
+     */
+    public $url;
+
+    /**
      * The context.
      *
      * @var Context
@@ -118,32 +123,5 @@ class Container extends Template
         $accessToken = $this->config->getAccessToken($storeId);
 
         return (string) $accessToken;
-    }
-
-    /**
-     * Get the page to navigate to within the protect client
-     *
-     * @return string The name of the page to naviage to.
-     */
-    public function getPageFromRequest(): string
-    {
-        $page = (string)$this->request->getParam('page');
-        $orderIncrementId = $this->getOrderIncrementIdFromRequest();
-        if (empty($page) && !empty($orderIncrementId)) {
-            $page = ClientSdkClient::CLIENT_PAGE_ORDER_DETAILS;
-        }
-
-        return $page;
-    }
-
-    /**
-     * Get the URL of the iframe that holds the NS8 Protect client.
-     *
-     * @return string The URL
-     */
-    public function getOrderIncrementIdFromRequest(): string
-    {
-        $orderId = $this->request->getParam('order_id');
-        return $orderId ? $this->order->getOrderIncrementId($orderId) : '';
     }
 }
